+++ CHANGES.rst
@@ -37,8 +37,10 @@ New Features
   * Add support for ``include_package_data`` and ``exclude_package_data`` setup keywords as well as parsing of
     ``MANIFEST.in``. See :issue:`315`. Thanks :user:`reiver-dev` for reporting the issue.
 
-  * Add support for ``cmake_minimum_required_version`` setup keyword.  See :issue:`312`.
+  * Add support for ``cmake_minimum_required_version`` setup keyword. See :issue:`312`.
     Suggested by :user:`henryiii`.
+
+  * Install cmake if found in ``setup_requires`` list. See :issue:`313`. Suggested by :user:`henryiii`.
 
 * Add support for ``--cmake-executable`` scikit-build command line option. Thanks :user:`henryborchers` for the suggestion.
   See :issue:`317`.